Revert "Miscellaneous: shut up warnings and a few fixes"
This reverts commit e7c2ac35508e076c85d5701ed43cf39b9b3cdbeb.

+++ camera/OMXCameraAdapter/OMXZoom.cpp
@@ -98,7 +98,7 @@ status_t OMXCameraAdapter::doZoom(int index)
         ret = -EINVAL;
         }
 
-    if (mPreviousZoomIndx == (unsigned int)index )
+    if (mPreviousZoomIndx == index )
         {
         return NO_ERROR;
         }
